Add a dependency on GNATCOLL
This also add LibGPR, as GNATCOLL requires it. Note however that this
temporarily does not add XML/Ada (which LibGPR requires) as we can't use
a fresh clone of it without running its configure script. This is
temporary, as the plan is to patch XML/Ada to support this use case.

For #140