Add Query.Options and Query.Native_Policies
The existing Query_Policies as been renamed to the more precise
Age_Policies, and a new Hint_Policies is added to select how unav-
ailable native releases are handled. By default, the new hinting
method will be used, in which missing native releases don't result
in an error but in a hint for external installation.

Both policies are passed to the solver inside a Query.Options record.