Refactor Supports_Hashing as a member function
This is done to allow a Filesystem origin to report True/False dep-
ending on its two use cases: deploying a local tarball (True) or
a local folder (False) -- the later is a working copy that may change.

An alternative would be to separate Filesystem into two origin kinds,
but that is much more work for no other practical benefit than this.