Alire.Os_Lib.Subprocess refactoring
The main goal of this patch is to move all the sub-process APIs to use
arguments in the form of a vector of strings rather than a single string
that is later sliced into different arguments. The reason for this
change is to better handle white-spaces in arguments, in particular
file-system paths.

The patch also removes unused or duplicate sub-programs in
Alire.Os_Lib.Subprocess and Alr.Os_Lib.

Because of the change of actions.command format in the manifests, this
patch requires an update of some crates in the index.