Replace the alr_build.gpr aggregate project by environment variables
Instead of creating an aggregate project that contains the path to
dependencies' gpr files, and the list of gpr files to build for the
crate. This patch introduce a procedure that sets the GPR_PROJECT_PATH
environment variable and runs gprbuild one time for each of of gpr file
of the crate. Same goes for gprclean.

The keys/values of external scenario variable are also set in the
environment.

This patch also introduces a setenv command that prints the
GPR_PROJECT_PATH and external scenario variable. This can be used to set
a correct build environment before starting an IDE for instance.