Ability to use a local folder to fulfill a dependency (#439)
* New Alire.Externals.Softlink to support WIP crates

* Implement alr with --url to add/pin in one step

A dependency can now be added simultaneously with a target directory that
fulfils it. Minimal check of the folder existing is performed; the buildability
of the whole is now the responsibility of the user.

* Implement softlinks in Alr.Dependencies.States

The placeholder `linked` fulfilment is now properly implemented to
hold/load/save the information about a directory to be used to fulfill a
dependency.

* Implement softlinks in Alire.Solutions.Solution

Add a Alire.Solutions.Linking subprogram that includes a linked dependency into
a solution.

* Modify solver to use soflinks for dependencies

When a softlink for the dependency being solved exists in the starting
solution, that dependency can be considered fulfilled without further ado.

* Implement `alr pin --url` to pin existing dep

This is the delayed counterpart to `alr with --url`; i.e., users can add the
dependency normally with `alr with` and later decide to pin it to a folder with
`alr pin`, without needing to remove it completely first with `alr with --del`.

We present pinning to versions and to folders to users under the same idea of
pinning. Internally, alr distinguishes between pinning (to versions) and
linking (to URLs). Currently only local paths can be used for links, but in the
future we may add remote files to fetch or repositories to clone.

* New Alire.URI to centralize URL knowledge

* Include pinned dirs' paths in project paths

The logic of collecting all necessary paths has been moved from Alr.Build_Env
to Alire.Roots.Project_Paths. A Root was already necessary to collect the
paths, and furthermore the root has all the necessary information: the root
release and the complete solution, which includes releases and linked dirs.

To make Roots self-contained, the platform properties are copied during startup
to Alire.Root. This is a temporary measure until these properties are
refactored from Alr into Alire.

* New tests for within/pinning to folders

* Testsuite fixes for minor changes in output

Information shown for external (lacking a release) dependencies is more
comprehensive now, showing transitivity, pin information, hinting status.

* Crate log of user-visible changes

* Spelling fixes

* Review: rename --url to --use