implementation of an automatic gpr file with (#458)
* First prototype implementation of an automatic gpr file with

* Add user query and config option for Auto-gpr-with

* Introduce Boolean properties

* Alire.Properties.Labeled: check TOML_Kind for labels

* auto-gpr-with: add crate entry to disable auto-gpr-with feature

* testsuite: allow custom indexes to reference fixtures/crates/

This removes duplication of several sources and archives.

* Bump index version to 0.3 for auto-gpr-with feature

* auto-gpr-with: do not add default project file for extern crates

* Bump alire-index in CI scripts

* tests/workflows/auto-gpr-with: rename test crate to avoid conflicts with scripts

* Review fixes

* tests/get/external-tool-dependency: make the test more robust

On Windows when using msys2, the make seen by alr can be different than
the one seen from the testsuite scripts. The means that the current way
to identify make's version will fail in some cases.

This patch introduces a different way to workaround the different
version numbers of make.