New Alire.Roots.Optional instead of Is_Valid
The Alire.Roots.Root now is always a root, that may exist or not on disk yet
(checked with Alire.Roots.Storage_Error). For the case where we may have a root
or not, and Outcome wrapper over Root is defined in Alire.Roots.Optional.

The change means that where before one would check a valid root with
Root.Current.Valid, now it is done with Root.Current.Exists.