Incrementally edit manifest on `alr with` (#484)
* Append new dependencies to the end of the manifest

* Conservative removal of dependencies in `alr with`

This action will also not touch lines that aren't involved in the deletion.
Only a simple dependency, and trivially empty [[depends-on]] array entries will
be removed. As an improvement over the previous situation, dynamic expressions
will not preclude using `alr with` for static dependencies, leaving the dynamic
ones untouched.

* Use lexicographical order for printing dependencies

This was happening before by chance in several tests. The `alr with` changes
have surfaced that the order on file is kept in tree; while this has no
secondary effects, for presentation to the user is better to be always
consistent. Solutions already use ordered sets, so for lists within conditional
trees we do the same (only for presentation).

* Testsuite minor output fix

* Tests for new removal of dependencies