New Alire.Roots.Optional type (#492)
In most cases, the Root has to be valid so it can be used. This patch moves the
Is_Valid code smell to a separate child type (Alire.Roots.Optional.Root) that
can be used in the few cases where all the root states have to be treated
separately.

This obsoletes Alr.Bootstrap.Session_States too.