Use GitHub API for basic checks/info during publishing (#554)
* Use GitHub API for basic checks/info during publish

* Code review fix

* Fix: new config values unavailable during same run

New config entries were stored on disk but not added to the memory map. As the
Config.Set subprogram isn't aware of the level of configuration being set,
there is no direct way to both store to disk and to memory; instead the
solution is that the configuration is reloaded every time. This is a bit of
overkill and a candidate for later optimization.