Add scenario to enable Ada 2020 features
I was trying to use the new generalized 'Image feature to debug, so this is a
convenient way of having Ada2020 features available without editing the gpr
file. Unfortunately, whatever is enabled by this switch breaks our parser of
manifests (spurious missing properties are reported), so it's not very useful
until that problem is pinpointed.