Alire.Environment: load env context from the platform  (#608)
* Alire.Environment: sort env list for predictable output

* Alire.Environment: load env context from the platform

For the moment this is only used on Windows for msys2. Before this patch the msys2 env was not included in `alr print` which means that the context was not complete.

* Alr.Platforms.Windows: do not install msys2 if pacman already in PATH

This makes alr usable when the user is already in an msys2 context.
And it alse makes possible to have an alire package in msys2.