Separate workflow to test unsupported package manager platform (#623)
* Workflow to test with unsupported package manager

* Add config option to disable distro detection

With test

* Fixes for circular elaboration

* Fix test for slight change in behavior

Before the last changes to break circularities, we tried to load the manifest
as part of the config loading process. Now we are satisfied with finding a
properly named file at the expected place. This causes one test to not detect
invalid TOML, since the manifest contents aren't really needed for the test.