Bump index version to devel-0.5 (#651)
* Bump index version to devel-0.5

The addition of more enum values to support other distributions is a
breaking change. This raises the question of whether we should simply drop
unknown enum values, as this could well work. But then, using an improper value
would silently fail. Perhaps the warning could be disabled for normal use and
re-enabled for publish?

* Fix python formatting (missing empty lines)