Remove Alire.Requisites and Alire.Boolean_Trees (#701)
* Availability as conditional tree

* Refactor up to Alire.Releases

* It builds for all of Alire.*

* It completely builds

* Testsuite passing, minor cosmetic fixes

There is a minor improvement for trees in which a vector with a single value is
promoted to the single value itself. As vectors are delimited by '(' ')', there
are a few test outputs that required eliminating these former extra
parentheses.

* Dead code removal related to old Ada index

* Case exprs excised from Requisites

This allows more dead code removal and some further internal simplifications.

* Removed Requisites and all related dead code

No longer needed, Requisites were useful in the old index for general
expressions, but not anymore in our case-restricted index. Some quite obscure
portions of Alire code can be thus removed.

* Bring gnatelim up to date

It bugboxes anyway, so it's not currently usable

* Stetic tweaks found during selfreview