Refactor the expression loader (#702)
* Pseudotypes specs

* Preparations to replace the generics mesh

* Buildable with the new design, still unimplemented

Only stub bodies

* New expression loader complete, pending ptypes

* Renamed packages Pseudotypes.* to Expressions.*

* Alire.Expressions.* implementation complete

* All functionality complete, tests missing

* System externals now reuse dynamic loaders

* Fixes found by the testsuite, and output tweaks

There is a functional difference that affects the output for some tests:
before, all enum cases received a value, even if it came from an "others"
clause. This meant that all cases were explicitly listed. Now, the "others" is
stored as is, and the output reproduces the contents of the TOML structure. A
few tests had to be adjusted for this.

* Fixes found testing the current stable index

* Remove obsolete documentation

* Implement and test 'x|y' expressions

* Improve output of `alr show` for system packages

* Fixes found during self-review: comments, dead code

* Fix for system externals report

* Alternative implementation for older GNATs