Transitively add pinned dirs (#713)
* Store relative pins when given as such

* Ensure pinned dirs use portable paths when relative

* Adjustments in testsuite for new relative links

* Recursively add links when linking crates

* Compound relative paths when adding links

* Restore inadvertently removed comment

* Relax too restrictive checks

In the case of linked releases, it doesn't matter if these contain expressions
as we don't write them to disk. As for the spurious warning during loading of
links, we may be transitively loading and the assumption on the current
directory is faulty in that case. The check must be performed elsewhere.

* Require approval for links to non-existent dirs

* Show broken links in dependency image

* Testsuite: new test and minimal tweaks

* Code review fixes

* Softlinks: report absolute project paths