Pins in the manifest (#743)
* Added data structures

* Pins may appear in manifest, being ignored

* Loading of user pins complete

* Pins are downloaded or skipped as needed

* Pins are properly pruned, and info displayed

Minor testsuite tweak for a change in logging format

* Added new-format pins to alire.toml for self-build

* Allow selective update of pins as for regular deps

* Local pins work with new manifest syntax

* Fixed bug in which version pins were not used

* Make version explicit key in user pin

* Fix bug about confirming empty updates

* Roots: Conflated dep updating into single Sync

We had two confusing Update_Dependencies and Update_And_Deploy_Dependencies
that were in practice doing almost the same. There is now a single
Sync_Dependencies.

* Disable tests that rely on `alr pin`

For now, these cannot work as we are going to remove the ability to edit pins
via `alr with`/`alr pin`. This functionality could be reintroduced at a later
time.

* More tests temporarily disabled

Most of those should be reimplemented in their manual edition alternative

* Disable code for pin edition in command-line

* Changes to allow pins to non-dependencies

This is purely for user comfort and will probably result in dependencies having
to be added at publish time. However, if we manage to restore command-line
pinning, we can remove that pain by adding missing dependencies at that time.

* Tests in tests/pin using new pins

Some tests are not easily portable without support from `alr with --use`. Since
that should be easy to implement later, they are disabled for now and will be
enabled in a subsequent patch.

* Documentation on new pins

* Fix for testcase on Windows using path separators

* Spelling fixes, and exclude lockfiles from check

* New test for various invalid pin entries in manifest

Also the required code changes to pass the test

* Fixes from code self-review

* Fixes suggested during code review

* Fix for missed update when there is no lockfile

We were creating an empty lockfile, which was newer than the manifest, and thus
not triggering the expected automatic update.

* Test to check pins are applied with no prior lockfile

This was a bug detected and corrected in the previous commit