Alr.Commands.Toolchain: tweak for multiple components and install dir (#822)
* Alr.Commands.Toolchain: tweak for multiple components and install dir

This patch adds support of multiple components and allow
--disable-assistant with all commands so that it is possible to
configure all the toolchain stuff in one command:
$ alr toolchain --select gnat_native gprbuild --disable-assistant

The patch also introduces a --install-dir switch to allow installation
in any directory. This can be useful for people willing to use the Alire
toolchain outside of an Alire workflow. It would be better if the root
dir of the archive (e.g. gprbuild_21.0.0_349f6f95) was skipped so that
all the tools would be available from <INSTALL_DIR>/bin/.

* tests: increase gnat_native version in fixture index and add more tests

* Alr.Commands.Toolchain: print list only when no command is specified

* tests/toolchain/custom_install_dir: fix paths for Windows and second dir

* Fix PR comments