prefixes with own nested root
We treat a global installation prefix simply as a workspace that is lastly
installed as a whole. In this way there cannot be inconsistencies with
libraries or multiple versions of anything. This kind of prefix only works for
now with crates coming from an index (so can be used as dependencies).

We also allow the installation of a work-in-progress workspace to a local
prefix, but this is mostly calling gprinstall without more ado.