Alire.Crate.Configuration: disable style checks in generated code
With the current style checks it is easy to break coding style by having
to many values in an enum for instance.

In theory we could generate better code for enums. But since users can
use their own coding style, it is impossible to be compatible with in
every situation.

So the best option is to disable all checks for this file.