Generate and use a providers cache file
This cache file, which is common to all indexes, tells us in advance which
crates have to be loaded because they provide some another crate. E.g., if
"gnat" is requested to be loaded, we need to load all gnat_native, and
x-compilers.

If this file doesn't exist, we can only be sure by loading the full index, at
which time the file is regenerated for future use.