Remove blue background setting from Windows shortcut (#1555)
The shortcut from the Windows installer was setting a background color for PowerShell command prompt. The idea was to match the look of the PowerShell you would normally start, and avoid confusion with the CMD prompt which has a black background. But that didn't work very well as the background color just disappeared when running commands. Also it seems like Windows now uses a black background for the PowerShell anyway.