Alr.Commands.Search: change default behavior to include properties
This patch comes from a comment of a user looking for gnatpp in Alire.
gnatpp is part of the libabalang_tools crate and a 1.x `alr search gnatpp`
returns nothing. The reason that by default only the release name and
descriptions are searched. This is quite a bad user interaction.

Here we change the default behavior to search in all properties of a
release, including tags, executables, website, etc.

We also display the list of properties that match the pattern.

Searching in names and descriptions only is still possible with the
--crates switch.