Rename option arguments to indicate expected type
In commands with options that admit arguments, like '--prefix' in
'alr install', is useful to provide an argument name in the command
help that gives a hint of the expected type.

For example, 'alr install -h' shows the message:
====
OPTIONS
   --prefix=ARG   Override installation prefix (default is ...)
====

When changed to:
====
OPTIONS
   --prefix=DIR   Override installation prefix (default is ...)
====

the 'DIR' name of the argument immediately says what is expected in this
option. In cases where this is not so obvious, it helps even more.
For example, in 'alr build -h':
====
OPTIONS
   --profiles=LIST     Comma-separated list of <crate>=<profile> values (see description)
   --stop-after=STAGE  Build stage after which to stop (see description)
====

is a quick remainder of what is expected in that options.